# pkg/policy/rule: Optimize rule String()