Avoiding to use  fmt.Sprintf() so that Go won't over-allocate the memory.

issue reference : cilium#19571

Result:

```
$ go test -v -run '^$' -bench 'BenchmarkRuleString' -benchmem ./pkg/policy > BenchmarkRuleString_old.txt
$ go test -v -run '^$' -bench 'BenchmarkRuleString' -benchmem ./pkg/policy > BenchmarkRuleString_new.txt
$ benchcmp BenchmarkRuleString_old.txt BenchmarkRuleString_new.txt
benchmark                 old ns/op     new ns/op     delta
BenchmarkRuleString-4     1929          1514          -21.51%

benchmark                 old allocs     new allocs     delta
BenchmarkRuleString-4     13             10             -23.08%

benchmark                 old bytes     new bytes     delta
BenchmarkRuleString-4     752           656           -12.77%
```

Signed-off-by: MikeLing <sabergeass@gmail.com>