@MikeLing 2022-05-17
> Checkpatch job seems to be failing because the commit msg author emails differ from the sign-off inside the commit. Could you fix those up?
> 
> ```
> Error: ERROR:NO_AUTHOR_SIGN_OFF: Missing Signed-off-by: line by nominal patch author 'linweijie08 <linweijie08@meituan.com>'
> ```

Already fix it. BTW, could I trigger the `tests` with `/test` by myself?